rename gtfs_rt_utils to model_update
since now this module has only one public method, we name it after this
method
