+++ src/lib.rs
@@ -11,7 +11,7 @@ pub(crate) mod utils;
 
 pub mod actors;
 pub mod context;
-pub(crate) mod gtfs_rt_utils;
+pub(crate) mod model_update;
 pub(crate) mod routes;
 pub mod server;
 pub mod siri_model;
